requirements.txt
dynaconf
smtplib
csv
email
python-tk
pillow